Solución a problema al mostrar los usuarios host con sus atributos - User

+++ api/src/main/java/com/hosting/rest/api/services/User/UserServiceImpl.java
@@ -1,5 +1,8 @@
 package com.hosting.rest.api.services.User;
 
+import static com.hosting.rest.api.Utils.AppUtils.isIntegerValidAndPositive;
+import static com.hosting.rest.api.Utils.AppUtils.isNotNull;
+
 import java.util.List;
 
 import javax.persistence.EntityManager;
@@ -9,7 +12,8 @@ import javax.persistence.TypedQuery;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.hosting.rest.api.exceptions.User.UserNotFoundException;
+import com.hosting.rest.api.exceptions.User.IllegalArgument.IllegalUserArgumentsException;
+import com.hosting.rest.api.exceptions.User.NotFound.UserNotFoundException;
 import com.hosting.rest.api.models.User.UserModel;
 import com.hosting.rest.api.repositories.User.IUserRepository;
 
@@ -28,15 +32,18 @@ public class UserServiceImpl implements IUserService {
 	}
 
 	@Override
-	public UserModel getUserById(Integer userId) {
-		return userRepo.findById(userId).orElseThrow(() -> new UserNotFoundException(userId));
+	public UserModel getUserById(final Integer userId) {
+		if (!isIntegerValidAndPositive(userId)) {
+			throw new IllegalUserArgumentsException("El id de usuario [ " + userId + " ] no es válido o está vacío.");
+		}
+
+		return userRepo.findById(userId)
+				.orElseThrow(() -> new UserNotFoundException("El usuario con id [ " + userId + " ] no existe."));
 	}
 
 	@Override
 	public List<UserModel> findAllUsers() {
-		// TODO:
-		String findAllHostUsersQuery = "select um" + " from UserModel um, UserHostModel uhm"
-				+ " where um.id = uhm.id";
+		String findAllHostUsersQuery = "SELECT um " + " FROM UserModel um" + " WHERE TYPE(um) IN (UserModel, UserHostModel)";
 
 		TypedQuery<UserModel> allUsers = em.createQuery(findAllHostUsersQuery, UserModel.class);
 
@@ -44,19 +51,63 @@ public class UserServiceImpl implements IUserService {
 	}
 
 	@Override
-	public UserModel addNewUser(UserModel userToAdd) {
+	public UserModel addNewUser(final UserModel userToAdd) {
+
+		if (!isNotNull(userToAdd)) {
+			throw new IllegalUserArgumentsException("Alguna propiedad del usuario a crear falta o no es válida.");
+		}
+
+		boolean existsUser = userRepo.findById(userToAdd.getId()).get() != null;
+
+		if (existsUser) {
+			throw new IllegalUserArgumentsException("Ya existe un usuario en la aplicación.");
+		}
+
 		return userRepo.save(userToAdd);
 	}
 
 	@Override
-	public UserModel updateUser(Integer userId, UserModel userToUpdate) {
-		return null;
-//		return userRepo.save(originalUser.get());
+	public UserModel updateUser(final Integer userId, final UserModel userToUpdate) {
+
+		if (!isNotNull(userToUpdate)) {
+			throw new IllegalUserArgumentsException("Alguna propiedad del usuario a actualizar falta o no es válida.");
+		}
+
+		boolean existsUser = userRepo.findById(userToUpdate.getId()).get() != null;
+
+		if (!existsUser) {
+			throw new IllegalUserArgumentsException("El usuario con id [ " + userId + " ] no existe.");
+		}
+
+		UserModel originalUser = userRepo.findById(userId).get();
+
+		// Nombre del usuario
+		originalUser.setName(userToUpdate.getName());
+
+		// Apellidos del usuario
+		originalUser.setSurname(userToUpdate.getSurname());
+
+		// Número de teléfono
+		originalUser.setSurname(userToUpdate.getSurname());
+
+		// Email
+		originalUser.setEmail(userToUpdate.getEmail());
+
+		return originalUser != null ? userRepo.save(originalUser) : null;
 	}
 
 	@Override
-	public void deleteUserById(Integer userId) {
+	public void deleteUserById(final Integer userId) {
+		if (!isIntegerValidAndPositive(userId)) {
+			throw new IllegalUserArgumentsException("Alguna propiedad del usuario a crear falta o no es válida.");
+		}
+
+		boolean existsUser = userRepo.findById(userId).get() != null;
+
+		if (!existsUser) {
+			throw new UserNotFoundException("No existe un usuario con id [ " + userId + " ]");
+		}
+
 		userRepo.deleteById(userId);
 	}
-
 }
